Refactor code in utils
