+++ test/cli-commands/utils/timeout.js
@@ -1,0 +1,7 @@
+function timeout(ms) {
+	return new Promise(resolve => setTimeout(resolve, ms));
+}
+
+module.exports = {
+	timeout
+};
